Move files to hw and change docker usage
This should have probably been several commits but it is what it is.
Moved most of the files to 'hw' folder to prepare for further addition
of 'sw' and other folders.
Besides that, rewrote most of the docker stuff in the makefile and
changed the former docker-compose.yml (now compose.yml) and Dockerfile
to accompany it.